# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "档案管理系统 API",
        "title": "depositum API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/libraries": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library"
                ],
                "summary": "获取库列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.PaginationResponse-github_com_redish101_depositum_pkg_api_v1_Library"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library"
                ],
                "summary": "创建库",
                "parameters": [
                    {
                        "description": "库",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.CreateLibraryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.Library"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/libraries/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library"
                ],
                "summary": "获取库详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "库 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.Library"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library"
                ],
                "summary": "删除库",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "库 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Library"
                ],
                "summary": "更新库",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "库 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "库",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.UpdateLibraryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.Library"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_redish101_depositum_pkg_api_v1.CreateLibraryRequest": {
            "type": "object",
            "required": [
                "address",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_redish101_depositum_pkg_api_v1.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_redish101_depositum_pkg_api_v1.Library": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_redish101_depositum_pkg_api_v1.PaginationResponse-github_com_redish101_depositum_pkg_api_v1_Library": {
            "type": "object",
            "properties": {
                "hasNext": {
                    "description": "是否有下一页",
                    "type": "boolean"
                },
                "hasPrev": {
                    "description": "是否有上一页",
                    "type": "boolean"
                },
                "items": {
                    "description": "数据列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_redish101_depositum_pkg_api_v1.Library"
                    }
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer"
                },
                "totalPages": {
                    "description": "总页数",
                    "type": "integer"
                }
            }
        },
...
        "github_com_redish101_depositum_pkg_api_v1.UpdateLibraryRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}